chore : inteface를 JSDoc 함수 뒤로 이동
