+++ src/apis/blocks/draftTextBlock.ts
@@ -1,14 +1,6 @@
 import { type ITextBlock } from '@/models/block'
 import { backAuthAPI } from '../api'
 
-export interface RequestTextBlockAddData {
-  content: string
-  x: number
-  y: number
-  h: number
-  w: number
-}
-
 /**
  * @description 텍스트블록 생성 API
  * @see https://www.notion.so/16b3bea33270447fa51a4a06858daac9
@@ -25,12 +17,14 @@ export const addTextBlockAPI = async (pageId: string | undefined, body: RequestT
     data: response.data.response
   }
 }
-
-export interface RequestTextBlockEditData {
-  objectId: ITextBlock['objectId']
-  content?: string
-  visible?: boolean
+export interface RequestTextBlockAddData {
+  content: string
+  x: number
+  y: number
+  h: number
+  w: number
 }
+
 /**
  * @description 텍스트블록 수정 API
  * @see https://www.notion.so/2-cdf1976fd3e641dd84bd77df574fb471?p=de99c95f224848148558741b65871dfa&pm=s
@@ -45,4 +39,9 @@ export const editTextBlockAPI = async (pageId: string | undefined, data: Request
     message: '텍스트블록을 수정했습니다.',
     data: res.response
   }
+}
+export interface RequestTextBlockEditData {
+  objectId: ITextBlock['objectId']
+  content?: string
+  visible?: boolean
 }
