+++ src/apis/blocks/imageBlock.ts
@@ -1,6 +1,27 @@
 import type { ImageBlock, EmbedGoogleMapBlock } from '@/models/block'
+import { JSONToForm } from '@/utils'
 import { backAuthAPI } from '../api'
 
+/**
+ * @description 이미지 블럭 생성 API
+ * @see https://www.notion.so/2ef9f22aad1a4836ad75df5446826013
+ */
+export const addImageBlockAPI = async (pageId: string | undefined, body: AddImageBlockBody) => {
+  if (!pageId) throw new Error('pageId가 없습니다.')
+  const form = JSONToForm(body)
+  const { data } = await backAuthAPI<AddImageBlockResponse>(`/v1/page/${pageId}/imageBlock`, {
+    method: 'POST',
+    data: form,
+    headers: {
+      'Content-Type': 'multipart/form-data'
+    }
+  })
+  return {
+    data: data.response,
+    status: 'success',
+    message: '이미지 블록을 추가했습니다.'
+  }
+}
 interface AddImageBlockResponse {
   response: EmbedGoogleMapBlock
   status: number
@@ -15,28 +36,23 @@ export interface AddImageBlockBody {
   title: ImageBlock['title']
   attachedImage: Blob
 }
-/**
- * @description 이미지 블럭 생성 API
- * @see https://www.notion.so/2ef9f22aad1a4836ad75df5446826013
- */
-export const addImageBlockAPI = async (pageId: string | undefined, body: AddImageBlockBody) => {
-  if (!pageId) throw new Error('pageId가 없습니다.')
-  const form = convertBodyToForm(body)
 
-  const { data } = await backAuthAPI<AddImageBlockResponse>(`/v1/page/${pageId}/imageBlock`, {
-    method: 'POST',
-    data: form,
-    headers: {
-      'Content-Type': 'multipart/form-data'
-    }
+/**
+ * @description 이미지 블럭 수정 API
+ * @see https://www.notion.so/73170cb13eba4ac5b5f34ed4ff113214
+ */
+export const editImageBlockAPI = async (pageId: string | undefined, body: EditImageBlockBody) => {
+  const form = JSONToForm(body)
+  const { data } = await backAuthAPI<EditImageBlockResponse>(`/v1/page/${pageId}/imageBlock/${body.objectId}`, {
+    method: 'PUT',
+    data: form
   })
   return {
     data: data.response,
     status: 'success',
-    message: '이미지 블록을 추가했습니다.'
+    message: '이미지 블록을 수정했습니다.'
   }
 }
-
 interface EditImageBlockResponse {
   response: ImageBlock
   status: number
@@ -47,22 +63,6 @@ export interface EditImageBlockBody {
   title?: ImageBlock['title']
   attachedImage?: Blob
   visible?: ImageBlock['visible']
-}
-/**
- * @description 이미지 블럭 수정 API
- * @see https://www.notion.so/73170cb13eba4ac5b5f34ed4ff113214
- */
-export const editImageBlockAPI = async (pageId: string | undefined, body: EditImageBlockBody) => {
-  const form = convertBodyToForm(body)
-  const { data } = await backAuthAPI<EditImageBlockResponse>(`/v1/page/${pageId}/imageBlock/${body.objectId}`, {
-    method: 'PUT',
-    data: form
-  })
-  return {
-    data: data.response,
-    status: 'success',
-    message: '이미지 블록을 수정했습니다.'
-  }
 }
 
 /**
@@ -79,20 +79,4 @@ export const deleteImageBlockAPI = async (pageId: string | undefined, blockId: I
     message: '이미지 블럭을 삭제했습니다.',
     data: response.data.response
   }
-}
-
-function convertBodyToForm<T>(body: T) {
-  type TKey = keyof T
-  const form = new FormData()
-  for (const key in body) {
-    const value = body[key as TKey]
-    if (typeof value === 'string') {
-      form.append(key, value)
-    } else if (value instanceof Blob) {
-      form.append(key, value, 'image.png')
-    } else {
-      form.append(key, String(value))
-    }
-  }
-  return form
 }
