+++ src/apis/blocks/linkBlock.ts
@@ -1,14 +1,6 @@
 import type { LinkBlock } from '@/models/block'
 import { backAuthAPI } from '../api'
 
-export type AddLinkBlockBody = {
-  x?: LinkBlock['x']
-  y?: LinkBlock['y']
-  w?: LinkBlock['w']
-  h?: LinkBlock['h']
-  title: LinkBlock['title']
-  link: LinkBlock['link']
-}
 /**
  * @description 링크 블럭 생성 API
  * @see https://www.notion.so/8bbf5e1b6e974bd89e78874644e320b7
@@ -24,13 +16,15 @@ export const addLinkBlockAPI = async (pageId: string | undefined, body: AddLinkB
     message: '링크 블록을 추가했습니다.'
   }
 }
-
-export type EditLinkBlockBody = {
-  objectId: LinkBlock['objectId']
-  title?: LinkBlock['title']
-  link?: LinkBlock['link']
-  visible?: LinkBlock['visible']
+export type AddLinkBlockBody = {
+  x?: LinkBlock['x']
+  y?: LinkBlock['y']
+  w?: LinkBlock['w']
+  h?: LinkBlock['h']
+  title: LinkBlock['title']
+  link: LinkBlock['link']
 }
+
 /**
  * @description 링크 블럭 수정 API
  * @see https:www.notion.so/107bc7366c3142f28b58b331989506fe
@@ -45,4 +39,10 @@ export const editLinkBlockAPI = async (pageId: string | undefined, body: EditLin
     status: 'success',
     message: '링크 블록을 수정했습니다.'
   }
+}
+export type EditLinkBlockBody = {
+  objectId: LinkBlock['objectId']
+  title?: LinkBlock['title']
+  link?: LinkBlock['link']
+  visible?: LinkBlock['visible']
 }
