+++ src/apis/blocks/videoBlock.ts
@@ -1,6 +1,27 @@
 import { type VideoBlock } from '@/models/block'
+import { JSONToForm } from '@/utils'
 import { backAuthAPI } from '../api'
 
+/**
+ * @description 비디오 블럭 생성 API
+ * @see https://www.notion.so/b3c240148be4470abb5ac0895de28575
+ */
+export const addVideoBlockAPI = async (pageId: string | undefined, body: AddVideoBlockBody) => {
+  if (!pageId) throw new Error('pageId가 없습니다.')
+  const form = JSONToForm(body)
+  const { data } = await backAuthAPI<AddVideoBlockResponse>(`/v1/page/${pageId}/videoBlock`, {
+    method: 'POST',
+    data: form,
+    headers: {
+      'Content-Type': 'multipart/form-data'
+    }
+  })
+  return {
+    data: data.response,
+    status: 'success',
+    message: '비디오 블록을 추가했습니다.'
+  }
+}
 interface AddVideoBlockResponse {
   response: VideoBlock
   status: number
@@ -15,27 +36,23 @@ export interface AddVideoBlockBody {
   videoLink?: string
   attachedVideo?: Blob
 }
+
 /**
- * @description 비디오 블럭 생성 API
- * @see https://www.notion.so/b3c240148be4470abb5ac0895de28575
+ * @description 비디오 블럭 수정 API
+ * @see https://www.notion.so/131f17790b8c4606891e4ca075f35f01
  */
-export const addVideoBlockAPI = async (pageId: string | undefined, body: AddVideoBlockBody) => {
-  if (!pageId) throw new Error('pageId가 없습니다.')
-  const form = convertBodyToForm(body)
-  const { data } = await backAuthAPI<AddVideoBlockResponse>(`/v1/page/${pageId}/videoBlock`, {
-    method: 'POST',
-    data: form,
-    headers: {
-      'Content-Type': 'multipart/form-data'
-    }
+export const editVideoBlockAPI = async (pageId: string | undefined, body: EditVideoBlockBody) => {
+  const form = JSONToForm(body)
+  const { data } = await backAuthAPI<EditVideoBlockResponse>(`/v1/page/${pageId}/videoBlock/${body.objectId}`, {
+    method: 'PUT',
+    data: form
   })
   return {
     data: data.response,
     status: 'success',
-    message: '비디오 블록을 추가했습니다.'
+    message: '비디오 블록을 수정했습니다.'
   }
 }
-
 interface EditVideoBlockResponse {
   response: VideoBlock
   status: number
@@ -47,22 +64,6 @@ export interface EditVideoBlockBody {
   attachedVideo?: Blob
   videoLink?: string
   visible?: VideoBlock['visible']
-}
-/**
- * @description 비디오 블럭 수정 API
- * @see https://www.notion.so/131f17790b8c4606891e4ca075f35f01
- */
-export const editVideoBlockAPI = async (pageId: string | undefined, body: EditVideoBlockBody) => {
-  const form = convertBodyToForm(body)
-  const { data } = await backAuthAPI<EditVideoBlockResponse>(`/v1/page/${pageId}/videoBlock/${body.objectId}`, {
-    method: 'PUT',
-    data: form
-  })
-  return {
-    data: data.response,
-    status: 'success',
-    message: '비디오 블록을 수정했습니다.'
-  }
 }
 
 /**
@@ -79,20 +80,4 @@ export const deleteVideoBlockAPI = async (pageId: string | undefined, blockId: V
     message: '비디오 블럭을 삭제했습니다.',
     data: response.data.response
   }
-}
-
-function convertBodyToForm<T>(body: T) {
-  type TKey = keyof T
-  const form = new FormData()
-  for (const key in body) {
-    const value = body[key as TKey]
-    if (typeof value === 'string') {
-      form.append(key, value)
-    } else if (value instanceof Blob) {
-      form.append(key, value, 'image.png')
-    } else {
-      form.append(key, String(value))
-    }
-  }
-  return form
 }
